push to live

+++ resources/views/admin/users/show.blade.php
@@ -36,7 +36,8 @@
                                         <th>{{ucfirst($attribute)}}</th>
                                         <?php
                                         if ($attribute == 'image'):
-                                            if (file_exists(public_path('uploads/users/' . $user->image))) {
+
+                                            if (file_exists(public_path('uploads\users\\' . $user->image)) && !empty($user->image)) {
                                                 echo "<td><img width='150' src=" . url('uploads/users/' . $user->image) . "></td>";
                                             } else {
                                                 echo "<td><img width='150' src=" . url('uploads/competition/noimage.png') . "></td>";
